+++ src/ts/easing.ts
@@ -1,4 +1,20 @@
-const easing: any = { // eslint-disable-line @typescript-eslint/no-explicit-any
+interface Easing {
+  [linear: string]: (t: number) => number
+  easeInQuad: (t: number) => number
+  easeOutQuad: (t: number) => number
+  easeInOutQuad: (t: number) => number
+  easeInCubic: (t: number) => number
+  easeOutCubic: (t: number) => number
+  easeInOutCubic: (t: number) => number
+  easeInQuart: (t: number) => number
+  easeOutQuart: (t: number) => number
+  easeInOutQuart: (t: number) => number
+  easeInQuint: (t: number) => number
+  easeOutQuint: (t: number) => number
+  easeInOutQuint: (t: number) => number
+}
+
+const easing: Easing = {
   linear (t: number): number {
     return t
   },